Add support for n-located conferences
In the same way as the program can determine in retrospect where a
conference should have been to minimize carbon costs, add a function to
determine how it should have been split.

Of course this is naive, since it does not take into account the fact
that the location of conferences influence who attends. Another thing
that could be done is to use as a dataset all the participants in the
history of the conference, only taking into account their most recent
departure city.